[da-vinci][server] Hybrid Quota Enforcer rename and improvements
- Renamed HybridStoreQuotaEnforcement to StorageUtilizationManager
  in order to better reflect its responsibilities, which are not
  only quota-related.

- The manager is now always initialized, rather than optional,
  which makes interactions with it more straightforward. The check
  of whether hybrid quota is enabled or not is evaluated inside
  the manager, rather than in the Store Ingestion Task.

- Reduced shared data structures between the Storage Utilization
  Manager and the Store Ingestion Task. In particular:

  - The map of StoragePartitionDiskUsage is now fully encapsulated
    and the Store Ingestion Task needs to go through the manager
    to populate or mutate its content. This eliminates some race
    condition in cases where the SIT is being invoked by multiple
    threads (e.g. in partition-wise shared consumer mode). It
    could also pave the way to swapping the implementation of the
    StoragePartitionDiskUsage to provide alternative update
    mechanisms or accuracy semantics. Left a TODO describing these.

  - The map of PartitionConsumptionState is now passed as an
    immutable collection, so the enforcer cannot put or remove
    entries in it. Unfortunately, both the enforcer and the Report
    Status Adapter mutate the PartitionConsumptionState entries
    themselves. This results in logic which is a bit tangled up,
    and it seems appropriate to improve that, but it would require
    a greater refactoring of the whole VeniceNotifier-related code.

- The Store Ingestion Task does not pass itself into the enforcer
  constructor anymore, and instead passes some constrained function
  handles for the specific actions the enforcer needs to invoke.

- The Store Ingestion Task does not delegate metric emissions to
  the manager anymore, but rather asks the manager for the current
  usage and then reports it on its own.

- Fixed a bug where the storage utilization value calculated for
  metric emission did not take amplification factor into account.

Unrelated:

- Tweaked StoreIngestionTask::produceToStoreBufferServiceOrKafka's
  signature to remove a bit of code duplication and better reflect
  the expectation that this function now handles a batch of events
  from just one partition.

- Eliminated a few redundant lookups of the Store object happening
  in various parts of the Store Ingestion Task's run loop. Instead,
  the Store is looked up once per loop iteration and passed around
  through the various functions that need it.

- Deleted the unused StorageQuotaExceededException.

RB=3296239
G=voldemort-reviewers
R=mnhuang,goliu,xhao,shuhliu,lcli,ssen,akozma,sumane,lelu,sdwu,anaberez,jlliu,yurzhou,zpolicze,xnma,smaji,nithakka,adietz,xyin
A=xhao

+++ da-vinci-client/src/main/java/com/linkedin/davinci/utils/StoragePartitionDiskUsage.java
@@ -35,7 +35,7 @@ public class StoragePartitionDiskUsage {
    * @param recordSize
    * @return true if recordSize >= 0
    */
-  public boolean add(int recordSize) {
+  public boolean add(long recordSize) {
     if (recordSize < 0) {
       return false;
     }
@@ -47,6 +47,15 @@ public class StoragePartitionDiskUsage {
    * When you want to check usage for a partition, use this method.
    * It will query syncing with real DB at a calculated frequency based on time and/or size trigger.
    * Generally calling this method should be quick
+   *
+   * TODO Consider alternative usage measurement mechanisms:
+   *      1. Register a listener to invocations of {@link AbstractStorageEngine#sync(int)} and refresh the disk
+   *         usage right after each sync, rather than doing it after some time lag.
+   *      2. Disregard the in-memory bookkeeping entirely, and only rely on the on-disk size, since the delta
+   *         should be bounded by the sync interval, and it is probably easier to reason about the system if it
+   *         can only pause consumption right after syncing to disk, rather than allowing the system to pause
+   *         in the middle of a sync interval, thus letting more data linger in-memory.
+   *
    * @return the disk usage for this partition
    */
   public long getUsage() {
